requirements.txt
pymupdf
loguru
langchain
langchain-openai
pillow
pandas
xlwings
openpyxl
PyPDF2
chardet
paddleocr
opencv-python
paddlepaddle
faiss-cpu
langchain_community
lark_oapi
requests_toolbelt
undetected_chromedriver
selenium
flask